updated composer libraries

+++ mediawiki/extensions/FacetedSearch2/src/Model/Response/Document.php
@@ -19,7 +19,7 @@ class Document
     public string $title;
     public string $displayTitle;
     public string $url;
-    public int $score;
+    public float $score;
     public ?string $highlighting;
 
     /**
@@ -32,10 +32,10 @@ class Document
      * @param string $title
      * @param string $displayTitle
      * @param string $url
-     * @param int $score
+     * @param float $score
      * @param string|null $highlighting
      */
-    public function __construct(string $id, array $propertyFacets, array $categoryFacets, array $directCategoryFacets, NamespaceFacetValue $namespaceFacet, string $title, string $displayTitle, string $url, int $score, ?string $highlighting)
+    public function __construct(string $id, array $propertyFacets, array $categoryFacets, array $directCategoryFacets, NamespaceFacetValue $namespaceFacet, string $title, string $displayTitle, string $url, float $score, ?string $highlighting)
     {
         $this->id = $id;
         $this->propertyFacets = $propertyFacets;
